Fix issue with `*` in move inputs
`*` was previously being escaped in move input while storing in FrameDb, since
it is a formatting character in Discord Markdown. However, this causes issues
when attempting to recognize move input queries. The input is now stored as is,
with the conversion to Discord Markdown done when the embed is being built.